with the cancel i can change a date from the editactivity

+++ app/src/main/java/it/polito/g13/dao/ReservationDao.kt
@@ -21,7 +21,7 @@ interface ReservationDao {
     fun getSingleReservation(id: Long) : Reservation
 
     @Query("SELECT * FROM $RESERVATION_TABLE where id == :id")
-    fun getASingleReservation(id: Long) : Reservation
+    fun getASingleReservation(id: Int) : Reservation
 
     @Query("SELECT * FROM $RESERVATION_TABLE")
     fun gettAllReservations() : LiveData<List<Reservation>>
